+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Support for sending GA4 [`add_payment_info`](https://developers.google.com/analytics/devguides/collection/ga4/reference/events#add_payment_info) when `vtex:addPaymentInfo` is received.
 - Support for sending GA4 [`begin_checkout`](https://developers.google.com/analytics/devguides/collection/ga4/reference/events#begin_checkout) when `vtex:beginCheckout` is received.
 - Support for sending GA4 [`view_cart`](https://developers.google.com/analytics/devguides/collection/ga4/reference/events#view_cart) when `vtex:beginCheckout` is received.
+- Support for sending GA4 [`refund`](https://developers.google.com/analytics/devguides/collection/ga4/reference/events#refund) when `vtex:refund` is received.
 - Support for sending GA4 [`add_to_wishlist`](https://developers.google.com/analytics/devguides/collection/ga4/reference/events#add_to_wishlist) when `vtex:addToWishlist` is received.
 
 ## [3.4.0] - 2023-02-15
