+++ react/__mocks__/transaction.ts
@@ -1,0 +1,55 @@
+const transactionProducts = [
+  {
+    attachments: [],
+    brand: 'New Offers!!',
+    brandId: '2000045',
+    category: 'Apparel & Accessories',
+    categoryId: '25',
+    categoryIdTree: ['25'],
+    categoryTree: ['Apparel & Accessories'],
+    components: [],
+    ean: '9812983',
+    id: '9',
+    measurementUnit: 'un',
+    name: 'Top Everyday Necessaire 100 RMS',
+    originalPrice: 1600.99,
+    price: 1600.99,
+    priceTags: [],
+    productRefId: '',
+    quantity: 1,
+    seller: 'VTEX',
+    sellerId: '1',
+    sellingPrice: 1600.99,
+    sku: '20',
+    skuName: '100 RMS',
+    skuRefId: '9812983',
+    slug: 'everyday-necessaire',
+    tax: 0,
+    unitMultiplier: 1,
+  },
+]
+
+export const transaction = {
+  accountName: 'storecomponents',
+  corporateName: null,
+  currency: 'USD',
+  event: '',
+  eventName: '',
+  openTextField: null,
+  orderGroup: '1310750551387',
+  ordersInOrderGroup: ['1310750551387-01'],
+  salesChannel: '1',
+  transactionAffiliation: 'VTEX',
+  transactionCurrency: 'USD',
+  transactionCustomTaxes: {},
+  transactionDate: '2023-02-14T18:56:47.0019167Z',
+  transactionDiscounts: 0,
+  transactionId: '1310750551387',
+  transactionLatestShippingEstimate: '2023-02-15T18:56:52.1493235Z',
+  transactionPayment: { id: 'FCE420A6B22C45D3BD60FD5DB55D34D1' },
+  transactionShipping: 1942.61,
+  transactionProducts,
+  transactionSubtotal: 1600.99,
+  transactionTax: 0,
+  transactionTotal: 3543.6,
+}
